[JIT] Handle del statements with variables as targets (#37608)
Summary:
**Summary**
This commit modifies the JIT frontend to handle `del` statements with
variables as targets by dropping the mapping corresponding to that
variable from the environment stack maintained by the IR emitter code.

**Test Plan**
This commit adds test cases for deleting a variable, deleting a variable
and then using it, and deleting a variable in a if-statement, and then
using it.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37608

Differential Revision: D21507239

Pulled By: SplitInfinity

fbshipit-source-id: ac7e353817dc76990ece294c95965cf585d6bdfb