[TensorPipe] Use PrefixStore to avoid conflicting keys (#39185)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/39185

The TP agent used the store for two things: mapping ranks to names, and mapping names to addresses. The former was prefixed, the latter wasn't. So, if a worker had a name which was `names/0` this would lead to a conflict. We should prefix both usages, and we can do so easily with the `PrefixStore`.
ghstack-source-id: 104837023

Test Plan: Unit tests

Differential Revision: D21767862

fbshipit-source-id: a256c0b9be349c7ffc11ac2790a2a682e3af84d5