[TensorPipe] Fix transport/channel priorities (#40090)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40090

I messed up in #39957: TensorPipe used to have a bug where it inverted priorities and preferred lower ones over higher ones. I had fixed that bug at the same time as I was writing that PR but then forgot to update the priority values once that PR landed. So this meant that TensorPipe was trying to bootstrap using SHM and then upgrade to UV. That worked in our tests because they are all run on the same machine, but that broke using TensorPipe across different machines. I'll take suggestions on how to have tests in place to prevent this type of breakages from happening.

The silver lining is that for some time our tests were testing the UV transport, instead of the SHM one, and it seems to be working alright. ;)
ghstack-source-id: 105967203

Differential Revision: D22067264

fbshipit-source-id: c6e3ae7a86038714cfba754b0811ca8a9a6f1347