Remove requirement of qnnpack engine for arm build. (#40112)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/40112

The changed code should be run for arm build even if qnnpack engine is not
enabled.
Furthermore the way AT_DISPATCH* stubs are defined, it just forms a lambda out
of the __VA__ARGS and executes the lambda. Thus return inside such lambda just
return to the original function and we end up executing the fallback path as
well.
Thus also changed #endif to #else...#endif.
This was causing per regression on mobile in one of the models.
ghstack-source-id: 105990691

Test Plan: CI

Reviewed By: supriyar

Differential Revision: D22072780

fbshipit-source-id: b12ca66aa19834b97b3eb0067af4e656cb8b3241